Fixed fingerprint overwrite bug
The request's session headers were being updated with X-Fingerprint *before* the rest of the headers were assigned, resulting in X-Fingerprint being overwritten. By simply assigning X-Fingerprint with the other headers instead of before, the bug is resolved.
---
I also fixed the strange capitalisation of the fingerprint variable to improve readability